requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.2
pycryptodome==3.10.1
python-decouple==3.4
pytz==2021.1
rave-python==1.2.13
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
